SASS stub, cleanup

+++ scripts/docs.sh
@@ -1,4 +1,11 @@
-#!/bin/sh
+#!/bin/bash
 
-rm -r docs/html
+set -e
+
+if [ -d "docs/html" ]; then
+    rm -r docs/html
+fi
+
+./scripts/styles.sh
+
 ./bin/sphinx-build docs/source/ docs/html
